requirements.txt
# Core ML & Deep Learning
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
transformers==5.5.0.dev0
accelerate==1.13.0
tokenizers==0.22.2
safetensors==0.7.0
sentencepiece==0.2.1
tiktoken==0.12.0

# Metric Learning & Training Utilities
pytorch-metric-learning==2.9.0
scikit-learn==1.8.0
scipy==1.17.1

# Data Processing
pandas==3.0.2
numpy==2.4.4
pyarrow==23.0.1
fastparquet==2026.3.0

# Visualization
matplotlib==3.10.8
pillow==12.1.1

# Experiment Tracking
comet_ml==3.57.3

# Config & Utilities
pyyaml==6.0.3
python-dotenv==1.2.2
tqdm==4.67.3
rich==14.3.3
typer==0.24.1
click==8.3.1

# Hugging Face Hub
huggingface-hub==1.8.0
fsspec==2026.3.0
hf-xet==1.4.3

# HTTP
requests==2.32.5
httpx==0.28.1
httpcore==1.0.9

# Misc
joblib==1.5.3
threadpoolctl==3.6.0
regex==2026.3.32
sympy==1.13.1
networkx==3.6.1
filelock==3.25.2
packaging==26.0
psutil==7.0.0
